requirements.txt
celery==5.0.5
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gevent==21.1.0
greenlet==1.0.0
python-dotenv==0.15.0
SQLAlchemy==1.3.22